Add setProjectionTransform for Platform
